censusindia: Census of India data for R

Digitised Census of India data from 1901 to 2011: population time series, primary census abstracts at state and district levels, mother tongue tables, SC/ST and tribal populations, and administrative directories. Census variables can be retrieved with geographic boundaries attached for mapping.

See also

Author

Maintainer: Saket Choudhary saketkc@gmail.com

Authors:

Saket Choudhary saketkc@gmail.com